backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
google-generativeai
python-dotenv
aiofiles
html2text
playwright